重构推测解码引擎：优化树结构、验证逻辑与基准测试
- 重构 tree.py 和 verification.py，改进候选树生成与验证流程
- 删除 judge.py 和 official_eval.py，精简评估逻辑
- 新增 tree15 多配置文件及对应流水线脚本
- 新增 test_integration_speedup 集成测试
- 更新 .gitignore 排除 logs/backups/.vscode 等非源码目录
- 调整 benchmark/engine/dflash_adapter 等核心模块

+++ src/dflash_specblock/rank_head.py
@@ -62,7 +62,7 @@ def valid_prefix_mask(logits: torch.Tensor, target_ids: torch.Tensor) -> torch.T
     correct = logits.argmax(dim=-1).eq(target_ids)
     prefix_before = torch.ones_like(correct, dtype=torch.bool)
     if correct.shape[-1] > 1:
-        prefix_before[..., 1:] = correct[..., :-1].cumprod(dim=-1).bool()
+        prefix_before[..., 1:] = correct[..., :-1].int().cumprod(dim=-1).bool()
     return prefix_before
